+++ visualpic/visualization/matplotlib/plot_containers/__init__.py
@@ -1,0 +1,5 @@
+from .figure import VPFigure
+from .subplots import FieldSubplot, ParticleSubplot
+
+
+__all__ = ['VPFigure', 'FieldSubplot', 'ParticleSubplot']
