+++ visualpic/visualization/matplotlib/plot_containers/figure.py
