+++ visualpic/visualization/matplotlib/plot_containers/subplots/__init__.py
@@ -1,0 +1,5 @@
+from .field_subplot import FieldSubplot
+from .particle_subplot import ParticleSubplot
+
+
+__all__ = ['FieldSubplot', 'ParticleSubplot']
