+++ visualpic/visualization/matplotlib/plot_types/__init__.py
@@ -1,0 +1,5 @@
+from .field_plot import field_plot
+from .particle_plot import particle_plot
+
+
+__all__ = ['field_plot', 'particle_plot']
